crfnet/requirements.txt
setuptools
numpy>=1.16.0
tensorflow-gpu==1.13.* 
tensorflow-serving-api-gpu==1.13.* 
opencv-python==3.* 
keras==2.2.4
pyquaternion 
h5py==2.7.1
Pillow 
progressbar2 
bayesian-optimization 
pandas 
pyarrow 
hilbertcurve
matplotlib
nuscenes-devkit @ git+https://github.com/Fellfalla/nuscenes-devkit.git@master
pydot
keras_resnet